+++ src/classes/sockets/primary-socket.ts
@@ -39,9 +39,6 @@ export default class PrimarySocket extends SocketBase {
         await GameSockets.emit(socketData);
         const timeKeeperSocketData = <SimpleCalendar.SimpleCalendarSocket.Data>{type: SocketTypes.clock, data: TimeKeeperStatus.Stopped};
         await GameSockets.emit(timeKeeperSocketData);
-        if(activeCalendar.time.unifyGameAndClockPause){
-            (<Game>game).togglePause(true, true);
-        }
         await MainApplication.timeKeepingCheck();
         Hook.emit(SimpleCalendarHooks.PrimaryGM, activeCalendar);
     }