+++ src/constants.ts
@@ -19,8 +19,11 @@ export enum SettingNames {
     OpenOnLoad = 'open-on-load',
     OpenCompact = 'open-compact',
     RememberPosition = 'remember-position',
+    RememberCompactPosition = 'remember-compact-position',
     AppPosition = 'app-position',
+    AppCompactPosition = 'app-compact-position',
     NoteReminderNotification = 'note-reminder-notification',
+    NoteListOpenDirection = 'note-list-open-direction',
     CalendarMainApp = 'calendar-main-app',
     CalendarConfigurationMenu= 'calendar-configuration-menu',
     CalendarConfiguration = 'calendar-configuration',
@@ -53,6 +56,8 @@ export const Themes: SimpleCalendar.Theme[] = [
     {key: 'wfrp4e', name: 'FSC.Configuration.Theme.WFRP4E', system: true, module: false},
     {key: 'forbidden-lands', name: 'FSC.Predefined.ForbiddenLands', system: true, module: false},
     {key: 'dsa5', name: 'FSC.Predefined.DSA-TDE5e', system: true, module: false},
+    {key: 'sfrpg', name: 'FSC.Predefined.Starfinder', system: true, module: false},
+    {key: 'wrath-and-glory', name: 'FSC.Predefined.Warhammer40kWG', system: true, module: false},
     //key: "ernies-modern-layout", name: "Ernie's Modern UI", system: false, module: true}
 ];