+++ CHANGELOG.md
@@ -1,9 +1,9 @@
 # Change Log
 
-## 2.1.34 - Theming, Bug Fixes & API Updates
+## 2.1.40 - Theming, Quality of Life Improvements, Bug Fixes & API Updates
 
 ![](https://img.shields.io/badge/release%20date-October%2030%2C%202022-blue)
-![GitHub release](https://img.shields.io/github/downloads-pre/vigoren/foundryvtt-simple-calendar/v2.1.34/module.zip)
+![GitHub release](https://img.shields.io/github/downloads-pre/vigoren/foundryvtt-simple-calendar/v2.1.40/module.zip)
 
 ### Theming
 
@@ -14,17 +14,30 @@ Changes that has been done to enable easy theming:
 - Changed all sizes to use REM instead of PX so that the entire interface can scale easily with any changes to the root REM size.
 - The addition of many CSS variables to quickly change how Simple Calendar looks. All the included themes use these variables to style SC.
 
-#### New Themes
+#### Themes
 
-To go along with these changes I have added a couple of new themes to Simple Calendar!
+To go along with these changes I have updated all the existing themes and added a couple of new themes to Simple Calendar!
 
-- A system specific theme for [Forbidden Lands](https://foundryvtt.com/packages/forbidden-lands).
+|                                                                                                                                                                                    |                                                                                                                                                                                                              |
+|------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------|--------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------|
+| Dark Theme ![](./docs/images/sc-theme-dark.png) ![](media://sc-theme-dark.png)                                                                                                     | Light Theme ![](./docs/images/sc-theme-light.png) ![](media://sc-theme-light.png)                                                                                                                            |
+| Classic Theme ![](./docs/images/sc-theme-classic.png) ![](media://sc-theme-classic.png)                                                                                            | [Warhammer Fantasy Roleplay 4th Edition](https://foundryvtt.com/packages/wfrp4e) System Theme ![](./docs/images/sc-theme-wfrp4e.png) ![](media://sc-theme-wfrp4e.png)                                        |
+| **NEW** [Forbidden Lands](https://foundryvtt.com/packages/forbidden-lands) System Theme  ![](./docs/images/sc-theme-forbidden-lands.png) ![](media://sc-theme-forbidden-lands.png) | **NEW** [Das Schwarze Auge / The Dark Eye (5th Edition)](https://foundryvtt.com/packages/dsa5) System Theme ![](./docs/images/sc-theme-dsa5.png) ![](media://sc-theme-dsa5.png)                              |
+| **NEW** [Starfinder](https://foundryvtt.com/packages/sfrpg) System Theme ![](./docs/images/sc-theme-sfrpg.png) ![](media://sc-theme-sfrpg.png)                                     | **NEW** [Warhammer 40,000 Roleplay: Wrath and Glory](https://foundryvtt.com/packages/wrath-and-glory) System Theme ![](./docs/images/sc-theme-wrath-and-glory.png) ![](media://sc-theme-wrath-and-glory.png) |
+
+
+### Quality of Life Improvements
+
+- Added a new client setting [Remember Compact Position](https://simplecalendar.info/pages/global-configuration/index/settings.html#client-settings). When enabled Simple Calendar will remember where the compact view is on the screen separate from where the full view is. This allows users to place the compact view in a corner, switch to full view and have the full view in the middle of the screen, then switch back to compact view, and it will return to the corner.
+- Added a new client setting [Side Drawer Open Direction](https://simplecalendar.info/pages/global-configuration/index/settings.html#client-settings). This allows users to choose which direction from the calendar side drawers like the note list will open. The current options are Right, Left and Down that will open the side drawers to the right, to the left and below the main calendar.
 
 ### Bug Fixes
 
 - Fixed a few instances where the notes dialog would not be sized correctly.
 - Fixed an issue where deleting a note after starting an edit would spawn several "Discard current changes" dialogs.
 - Fixed a bug where notes shown in the search result list would not properly show who had access to view the note.
+- Fixed a bug where the compact view would not have the correct height set.
+- Fixed a bug with the `Unify Clock Start/Pause With Game Pause` setting that would make it so the game always started paused. (Fixes a compatibility error with the [Koboldworkds - Pause Control](https://foundryvtt.com/packages/koboldworks-pause-control) module).
 
 ### API Updates
 
@@ -49,6 +62,11 @@ Thank you to the follow people for making updates to Simple Calendars translatio
 
 ### Documentation Changes
 
+- Added a link to the GitHub page for the module in the header of the documentation site.
+- Added documentation for [customizing the themes](https://simplecalendar.info/developing-with-sc/index/theming.html) of Simple Calendar.
+- Added documentation for [contributing](https://github.com/vigoren/foundryvtt-simple-calendar/blob/main/CONTRIBUTING.md#themes) a new theme to Simple Calendar.
+- Added documentation that outlines all the [available themes](https://simplecalendar.info/pages/docs/using-sc/index/themes.html) within Simple Calendar.
+- Updated the [client settings](https://simplecalendar.info/pages/global-configuration/index/settings.html#client-settings) documentation to include the new settings `Remember Compact Position` and `Side Drawer Open Direction`.
 - Fixed some typos in the example of the [setDate API function](https://simplecalendar.info/functions/SimpleCalendar.api.setDate.html)
 - Fixed typos in the API documentation when referencing a `second` property when it should be `seconds`.
 - Updated the example for the [timestampToDate](https://simplecalendar.info/functions/SimpleCalendar.api.timestampToDate.html) API function.